refactor: code separation
